Refactor and document interactive map component
Split the map component into two, one wrapper that is nicer to hack on
and talks to stores and actions, and a lower-level leaflet-oriented one
with the ugly DOM stuff.

Unfortunately, children of the map still have to pass a `getMap`
function down to the line function down to their descendents if they
want put leaflet layers on the map. I couldn't figure out how to
isolate that concern from other components.

This commit fixes both ASDB-53 and ASDB-53.