Refactor geoJson component and add custom icon
Added callbacks to the PointLayer component for custom pins, and a hook
into the double-click event.

The double-click event hook doesn't really generalize to any event we
might want to hook into in a nice way, but it might be good enough.

Double-click on an icon in the international view navigates to that
country.

This commit resolves ASDB-27.